Example to test multi-threaded progressbar handling
This will start a loop using a progressbar when both the left & right
mouse buttons are pressed together. When the if statement at
src/gui/mrview/window.cpp, line 1398 is true, the loop runs within the
current thread. If false, it runs in a separate thread. Importantly,
when run in a separate thread, the current thread will not block waiting
for it to finish - it connects to the thread's finished() signal to
execute a follow-on callback.

This means that if any functions are expected to run in multiple
threads while using a progressbar, the way to handle that is to
encapsulate them in a QThread (as in this example), and perform the rest
of the processing in a separate slot...